+++ src/@types/vmlog.zdump.d.ts
@@ -1,0 +1,8 @@
+type ZdumpInfo = {
+  build     : string,
+  panic_date: string,
+  panic_msg : string,
+  uptime    : string,
+  trace     : Array<string>,
+  log       : Array<string>
+}